requirements.txt
asyncio
websockets
mysql-connector-python